增加15位以内数字转中文的方法

+++ app/src/main/java/com/lodz/android/agiledevkt/modules/format/NumFormatTestActivity.kt
@@ -62,7 +62,7 @@ class NumFormatTestActivity : BaseActivity() {
                 toastShort(R.string.format_num_hint)
                 return@setOnClickListener
             }
-            mBinding.resultTv.text = str.toInt().toChinese()
+            mBinding.resultTv.text = str.toLong().toChinese()
         }
     }
